Add 'eanHandler' & move code from the method to it
